+++ cmd/marble/main.go
@@ -41,7 +41,6 @@ import (
 	neocompute "github.com/R3E-Network/service_layer/services/confcompute/marble"
 	neooracle "github.com/R3E-Network/service_layer/services/conforacle/marble"
 	neofeeds "github.com/R3E-Network/service_layer/services/datafeed/marble"
-	neorand "github.com/R3E-Network/service_layer/services/vrf/marble"
 	txproxy "github.com/R3E-Network/service_layer/services/txproxy/marble"
 )
 
@@ -60,7 +59,6 @@ var availableServices = []string{
 	"neofeeds",
 	"neoflow",
 	"neooracle",
-	"neorand",
 	"txproxy",
 }
 
@@ -194,15 +192,6 @@ func main() {
 			automationHash = trimHexPrefix(string(secret))
 		} else if secret, ok := m.Secret("CONTRACT_NEOFLOW_HASH"); ok && len(secret) > 0 {
 			automationHash = trimHexPrefix(string(secret))
-		}
-	}
-
-	randomnessLogHash := trimHexPrefix(contracts.RandomnessLog)
-	if randomnessLogHash == "" {
-		if secret, ok := m.Secret("CONTRACT_RANDOMNESSLOG_HASH"); ok && len(secret) > 0 {
-			randomnessLogHash = trimHexPrefix(string(secret))
-		} else if secret, ok := m.Secret("CONTRACT_RANDOMNESS_LOG_HASH"); ok && len(secret) > 0 {
-			randomnessLogHash = trimHexPrefix(string(secret))
 		}
 	}
 
@@ -349,15 +338,6 @@ func main() {
 			Marble:         m,
 			SecretProvider: newServiceSecretsProvider(m, db, neooracle.ServiceID),
 			URLAllowlist:   oracleAllowlist,
-		})
-	case "neorand":
-		svc, err = neorand.New(neorand.Config{
-			Marble:            m,
-			DB:                db,
-			GlobalSignerURL:   globalSignerURL,
-			ChainClient:       chainClient,
-			ChainSigner:       teeSigner,
-			RandomnessLogHash: randomnessLogHash,
 		})
 	case "txproxy":
 		svc, err = txproxy.New(txproxy.Config{